DS_Hackton06/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
jupyterlab
joblib
python-dotenv
kaggle
shap
optuna
pydantic
pillow
torch
torchvision
tqdm
requests
albumentations
opencv-python-headless
pytest
torchmetrics
